hw02-api-transport-models / tests added
