+++ ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/common/ErrorDto.kt
@@ -1,0 +1,20 @@
+package ru.otus.otuskotlin.propertysale.mp.transport.ps.common
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ErrorDto(
+    val code: String? = null,
+    val group: String? = null,
+    val field: String? = null,
+    val level: Level? = null,
+    val message: String? = null,
+) {
+    @Serializable
+    enum class Level {
+        SUCCESS,
+        INFO,
+        WARNING,
+        ERROR,
+    }
+}
