+++ ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/common/IPsDebug.kt
@@ -1,0 +1,5 @@
+package ru.otus.otuskotlin.propertysale.mp.transport.ps.common
+
+interface IPsDebug {
+    val mode: PsWorkModeDto?
+}
