+++ ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/common/IPsRequest.kt
@@ -1,0 +1,8 @@
+package ru.otus.otuskotlin.propertysale.mp.transport.ps.common
+
+interface IPsRequest {
+    val requestId: String?
+    val onResponse: String?
+    val startTime: String?
+    val debug: IPsDebug?
+}
