+++ ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/common/IPsResponse.kt
@@ -1,0 +1,10 @@
+package ru.otus.otuskotlin.propertysale.mp.transport.ps.common
+
+interface IPsResponse {
+    val responseId: String?
+    val onRequest: String?
+    val endTime: String?
+    val errors: List<ErrorDto>?
+    val status: ResponseStatusDto?
+    val debug: IPsDebug?
+}
