+++ ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/common/PsItemPermission.kt
@@ -1,0 +1,7 @@
+package ru.otus.otuskotlin.propertysale.mp.transport.ps.common
+
+enum class PsItemPermission {
+    READ,
+    UPDATE,
+    DELETE,
+}
