+++ ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/common/PsMessage.kt
@@ -1,0 +1,6 @@
+package ru.otus.otuskotlin.propertysale.mp.transport.ps.common
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+abstract class PsMessage()
