+++ ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/common/PsWorkModeDto.kt
@@ -1,0 +1,9 @@
+package ru.otus.otuskotlin.propertysale.mp.transport.ps.common
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class PsWorkModeDto {
+    PROD,
+    TEST
+}
