+++ ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/common/ResponseStatusDto.kt
@@ -1,0 +1,8 @@
+package ru.otus.otuskotlin.propertysale.mp.transport.ps.common
+
+enum class ResponseStatusDto {
+    SUCCESS,
+    BAD_REQUEST,
+    INTERNAL_SERVER_ERROR,
+    NOT_FOUND,
+}
