+++ ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/PsActionDto.kt
@@ -1,0 +1,11 @@
+package ru.otus.otuskotlin.propertysale.mp.transport.ps.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PsActionDto(
+    val id: String? = null,
+    val type: String? = null,
+    val content: String? = null,
+    val status: String? = null,
+)
