+++ ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/PsListFilterDto.kt
@@ -1,0 +1,8 @@
+package ru.otus.otuskotlin.propertysale.mp.transport.ps.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PsListFilterDto(
+    val text: String? = null,
+)
