+++ ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/PsUnitTypeDto.kt
@@ -1,0 +1,15 @@
+package ru.otus.otuskotlin.propertysale.mp.transport.ps.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PsUnitTypeDto(
+    val id: String? = null,
+    val name: String? = null,
+    val description: String? = null,
+    val synonyms: Set<String>? = null,
+    val symbol: String? = null,
+    val symbols: Set<String>? = null,
+    val isBase: Boolean? = null,
+)
+
