--- ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/common/IPsDebug.kt
@@ -1,5 +1,0 @@
-package ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common
-
-interface IPsDebug {
-    val mode: PsWorkModeDto?
-}
