--- ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/common/IPsRequest.kt
@@ -1,8 +1,0 @@
-package ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common
-
-interface IPsRequest {
-    val requestId: String?
-    val onResponse: String?
-    val startTime: String?
-    val debug: IPsDebug?
-}
