--- ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/common/PsItemPermission.kt
@@ -1,7 +1,0 @@
-package ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common
-
-enum class PsItemPermission {
-    READ,
-    UPDATE,
-    DELETE,
-}
