--- ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/common/PsMessage.kt
@@ -1,6 +1,0 @@
-package ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-abstract class PsMessage()
