--- ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/common/PsWorkModeDto.kt
@@ -1,9 +1,0 @@
-package ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-enum class PsWorkModeDto {
-    PROD,
-    TEST
-}
