--- ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/common/ResponseStatusDto.kt
@@ -1,8 +1,0 @@
-package ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common
-
-enum class ResponseStatusDto {
-    SUCCESS,
-    BAD_REQUEST,
-    INTERNAL_SERVER_ERROR,
-    NOT_FOUND,
-}
