--- ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/crud/PsResponseRead.kt
@@ -1,29 +1,0 @@
-package ru.otus.otuskotlin.propertysale.mp.transport.ps.models.crud
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common.ErrorDto
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common.IPsDebug
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common.IPsResponse
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common.PsMessage
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common.PsWorkModeDto
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.models.common.ResponseStatusDto
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.models.models.PsDto
-
-@Serializable
-@SerialName("PsResponseRead")
-data class PsResponseRead(
-    override val responseId: String? = null,
-    override val onRequest: String? = null,
-    override val endTime: String? = null,
-    override val debug: Debug? = null,
-    override val errors: List<ErrorDto>? = null,
-    override val status: ResponseStatusDto? = null,
-    val ps: PsDto? = null,
-) : IPsResponse, PsMessage() {
-
-    @Serializable
-    data class Debug(
-        override val mode: PsWorkModeDto?
-    ) : IPsDebug
-}
