--- ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/models/PsActionDto.kt
@@ -1,11 +1,0 @@
-package ru.otus.otuskotlin.propertysale.mp.transport.ps.models.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class PsActionDto(
-    val id: String? = null,
-    val type: String? = null,
-    val content: String? = null,
-    val status: String? = null,
-)
