--- ok-propertysale-mp-transport-ps/src/commonMain/kotlin/ru/otus/otuskotlin/propertysale/mp/transport/ps/models/models/PsListFilterDto.kt
@@ -1,8 +1,0 @@
-package ru.otus.otuskotlin.propertysale.mp.transport.ps.models.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class PsListFilterDto(
-    val text: String? = null,
-)
